Version 1.4.0

+++ stm32f7508_freertos-bsp/projects/microej/core/inc/microej.h
@@ -1,7 +1,7 @@
 /*
  * C
  *
- * Copyright 2013-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2013-2021 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
@@ -15,7 +15,7 @@
 
 /* Defines -------------------------------------------------------------------*/
 
-#define MICROEJ_FALSE 	0
-#define MICROEJ_TRUE 	1
+#define MICROEJ_FALSE 	((uint8_t) 0)
+#define MICROEJ_TRUE 	((uint8_t) 1)
 
 #endif
